agrointent/backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-cloud-logging
google-generativeai
python-multipart
pillow
slowapi